# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "epubapi",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/login": {
      "post": {
        "tags": [
          "crate::service::user::route"
        ],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id",
                  "password"
                ],
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "required": [
                        "unauthorized"
                      ],
                      "properties": {
                        "unauthorized": {
                          "type": "string"
                        }
                      }
                    },
                    {
                      "type": "object",
                      "required": [
                        "invalid id or password"
                      ],
                      "properties": {
                        "invalid id or password": {
                          "type": "string"
                        }
                      }
                    },
                    {
                      "type": "object",
                      "required": [
                        "invalid invitation code"
                      ],
                      "properties": {
                        "invalid invitation code": {
                          "type": "string"
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "invalid id or password": "invalid id or password"
                }
              }
            }
          }
        }
      }
    },
...
    "/users": {
      "get": {
        "tags": [
          "crate::service::user::route"
        ],
        "summary": "ユーザー情報を取得",
        "operationId": "show_user",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "role",
                    "api_key",
                    "invitations"
                  ],
                  "properties": {
                    "api_key": {
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    },
                    "invitations": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "role": {
                      "type": "string",
                      "enum": [
                        "Admin",
                        "User"
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "required": [
                        "unauthorized"
                      ],
                      "properties": {
                        "unauthorized": {
                          "type": "string"
                        }
                      }
                    },
                    {
                      "type": "object",
                      "required": [
                        "invalid id or password"
                      ],
                      "properties": {
                        "invalid id or password": {
                          "type": "string"
                        }
                      }
                    },
                    {
                      "type": "object",
                      "required": [
                        "invalid invitation code"
                      ],
                      "properties": {
                        "invalid invitation code": {
                          "type": "string"
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "unauthorized": "認証に失敗しました"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "crate::service::user::route"
        ],
        "operationId": "new_user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id",
                  "password",
                  "invitation_code"
                ],
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "invitation_code": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "required": [
                        "unauthorized"
                      ],
                      "properties": {
                        "unauthorized": {
                          "type": "string"
                        }
                      }
                    },
                    {
                      "type": "object",
                      "required": [
                        "invalid id or password"
                      ],
                      "properties": {
                        "invalid id or password": {
                          "type": "string"
                        }
                      }
                    },
                    {
                      "type": "object",
                      "required": [
                        "invalid invitation code"
                      ],
                      "properties": {
                        "invalid invitation code": {
                          "type": "string"
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "invalid id or password": "IDかパスワードが不正です"
                }
              }
            }
          }
        }
      }
    }
  },
...
}